+++ src/view/film-container.js
@@ -9,4 +9,10 @@ export default class FilmContainer extends Abstract {
   getTemplate() {
     return createFilmContainerTemplate();
   }
+
+  removeElement() {
+    this._element.remove();
+
+    super.removeElement();
+  }
 }
